# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Tutorial",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.laravel-swagger-tutorial.test/",
            "description": "API server"
        }
    ],
    "paths": {
        "/TestCrud": {
            "get": {
                "tags": [
                    "test crud"
                ],
                "summary": "List all test Crud",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "user"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An paged array of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TestCrudResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "New test crud create",
                "operationId": "store",
                "requestBody": {
                    "description": "Post object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/testCrudRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestCrudResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TestCrudResponse": {
                "title": "TestCrudResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "attributes": {
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "body": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "relationships": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "title": "UserResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "attributes": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "testCrudRequest": {
                "title": "testCrudRequest",
                "description": "Class PostViewModel",
                "required": [
                    "title",
                    "body"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}